refactor(server): moving methods and types
