+++ packages/amplication-server/src/core/auth/auth-utils.ts
@@ -1,0 +1,9 @@
+import * as crypto from "crypto";
+
+export function generateRandomString(): string {
+  return crypto.randomBytes(10).toString("hex");
+}
+
+export function generateRandomEmail(): string {
+  return `fake-${generateRandomString()}@amplication.com`;
+}
