+++ packages/amplication-server/src/core/workspace/workspace.module.ts
@@ -12,6 +12,7 @@ import { BillingModule } from "../billing/billing.module";
 import { ModuleModule } from "../module/module.module";
 import { ModuleActionModule } from "../moduleAction/moduleAction.module";
 import { WorkspaceController } from "./workspace.controller";
+import { ResourceModule } from "../resource/resource.module";
 
 @Module({
   imports: [
@@ -21,6 +22,7 @@ import { WorkspaceController } from "./workspace.controller";
     MailModule,
     UserModule,
     ProjectModule,
+    ResourceModule,
     SubscriptionModule,
     BillingModule,
     ModuleModule,
