+++ packages/amplication-server/src/core/workspace/workspace.service.ts
@@ -1,5 +1,13 @@
 import { ConflictException, Inject, Injectable } from "@nestjs/common";
-import { Entity, EntityField, Resource, User, Workspace } from "../../models";
+import {
+  Account,
+  Entity,
+  EntityField,
+  Project,
+  Resource,
+  User,
+  Workspace,
+} from "../../models";
 import { Prisma, PrismaService } from "../../prisma";
 import {
   CompleteInvitationArgs,
@@ -36,6 +44,7 @@ import { BillingPeriod } from "@stigg/node-server-sdk";
 import { Coupon } from "./dto/Coupon";
 import { EnumResourceType } from "../resource/dto/EnumResourceType";
 import {
+  EnumAuthProviderType,
   EnumBlockType,
   EnumDataType,
 } from "@amplication/code-gen-types/models";
@@ -47,8 +56,20 @@ import { BillingLimitationError } from "../../errors/BillingLimitationError";
 import { Env } from "../../env";
 import { ConfigService } from "@nestjs/config";
 import { EnumPreviewAccountType } from "../auth/dto/EnumPreviewAccountType";
+import { ResourceService } from "../resource/resource.service";
+import { generateRandomString } from "../auth/auth-utils";
+import { AuthUser, CreatePreviewServiceSettingsArgs } from "../auth/types";
+import { USER_ENTITY_NAME } from "../entity/constants";
 
 const INVITATION_EXPIRATION_DAYS = 7;
+
+type PreviewAccountEnvironment = {
+  workspace: Workspace & {
+    users: AuthUser[];
+  };
+  project: Project;
+  resource: Resource;
+};
 
 @Injectable()
 export class WorkspaceService {
@@ -59,6 +80,7 @@ export class WorkspaceService {
     private readonly mailService: MailService,
     private readonly subscriptionService: SubscriptionService,
     private readonly projectService: ProjectService,
+    private readonly resourceService: ResourceService,
     private readonly billingService: BillingService,
     private analytics: SegmentAnalyticsService,
     private readonly moduleService: ModuleService,
@@ -1109,5 +1131,123 @@ export class WorkspaceService {
     });
     await Promise.allSettled(promises);
     return hasChanges;
+  }
+
+  /**
+   * workspace, project and service creation for preview account
+   */
+
+  async createPreviewEnvironment(
+    account: Account
+  ): Promise<PreviewAccountEnvironment> {
+    const workspaceName = `Amplication-${generateRandomString()}`;
+    const projectName = account.previewAccountType;
+
+    const workspace = (await this.createPreviewWorkspace(
+      {
+        data: {
+          name: workspaceName,
+        },
+        include: {
+          users: {
+            include: {
+              account: true,
+              userRoles: true,
+              workspace: true,
+            },
+          },
+        },
+      },
+      account.id
+    )) as unknown as Workspace & { users: AuthUser[] };
+    const [user] = workspace.users as AuthUser[];
+
+    const project = await this.projectService.createProject(
+      {
+        data: {
+          name: projectName,
+          workspace: {
+            connect: {
+              id: workspace.id,
+            },
+          },
+        },
+      },
+      user.id
+    );
+
+    const resource = await this.createPreviewServiceWithPredefinedSettings(
+      project.id,
+      user
+    );
+
+    return {
+      workspace,
+      project,
+      resource,
+    };
+  }
+
+  private async createPreviewServiceWithPredefinedSettings(
+    projectId: string,
+    user: User
+  ): Promise<Resource> {
+    const serviceName = "Monolith";
+    const previewServiceSettings = this.createPreviewServiceSettings({
+      projectId: projectId,
+      name: serviceName,
+      description: "Monolith Service",
+      adminUIPath: "",
+      serverPath: `./apps/${serviceName}`,
+      generateAdminUI: false,
+      generateGraphQL: true,
+      generateRestApi: true,
+    });
+
+    const resource = await this.resourceService.createPreviewService({
+      args: previewServiceSettings,
+      user,
+      nonDefaultPluginsToInstall: [],
+      requireAuthenticationEntity: false,
+    });
+
+    return resource;
+  }
+
+  private createPreviewServiceSettings({
+    name,
+    description,
+    adminUIPath,
+    serverPath,
+    generateAdminUI,
+    generateGraphQL,
+    generateRestApi,
+    projectId,
+  }: CreatePreviewServiceSettingsArgs) {
+    return {
+      data: {
+        name,
+        description,
+        resourceType: EnumResourceType.Service,
+        project: {
+          connect: {
+            id: projectId,
+          },
+        },
+        serviceSettings: {
+          authProvider: EnumAuthProviderType.Jwt,
+          authEntityName: USER_ENTITY_NAME,
+          adminUISettings: {
+            adminUIPath,
+            generateAdminUI,
+          },
+          serverSettings: {
+            generateGraphQL,
+            generateRestApi,
+            serverPath,
+          },
+        },
+      },
+    };
   }
 }
